requirements.txt
requests>=2.32.0
PyYAML>=6.0.1
streamlit>=1.40.0
pandas>=2.2.0
plotly>=5.24.0
